// test/types/api.ts
export enum SubmitRoutingOrigin {
  SuccessOrigin = "SuccessOrigin",
  FailureOrigin = "FailureOrigin",
  InProgressOrigin = "InProgressOrigin",
  ProcessErrorOrigin = "ProcessErrorOrigin",
  ServerErrorOrigin = "ServerErrorOrigin",
}

export enum SubmitRoutingToken {
  SuccessToken = "SuccessToken",
  FailureToken = "FailureToken",
  InProgressToken = "InProgressToken",
  ProcessErrorToken = "ProcessErrorToken",
}